requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-filter==25.2
django-sslserver==0.22
django-taggit==6.1.0
djangorestframework==3.16.1
mysql-connector-python==9.4.0
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
